requirements.txt
streamlit
pandas
openpyxl
numpy
matplotlib
pyyaml
xlrd
xlwt
xlutils

